+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 * Fix ui and ux issues
 * Fix songs not displaying porperly on Mac and Linux
 * Fix selection and flyout tool not showing properly
+* Fix the look of timer input
 ---