+++ CHANGELOG.md
@@ -21,7 +21,7 @@ All notable changes to this project will be documented in this file.
  
  ---
 
- ## [4.0.4] - 2025-08-17
+ ## [4.0.4] - 2026-04-28
 ### Added
 * Added countdown timer to song line duration.
 * Added color gradient